requirements.txt
PyNaCl==1.5.0
discord.py==2.5.2
regex==2024.5.15
gTTS==2.5.0
pyaml==23.9.7
pydub==0.25.1
requests==2.31.0
zipfile36==0.1.3
feedparser==6.0.11
beautifulsoup4==4.12.3
fuzzywuzzy[speedup]==0.18.0
flask==3.1.0